Refactor EBA Template 4 models for EU 2023/2083 compliance
Addressed multiple audit findings (F-16 to F-20, F-28) related to Template 4.1
to strictly align with EBA Annex II standards:

- npl_portfolio/property_collateral_choices.py:
  Realigned TYPE_OF_PROPERTY_CHOICES (4.01) to the 5 EBA options and
  TYPE_OF_OCCUPANCY_CHOICES (4.17) to the 3 EBA options. Added legacy
  subtype choices to preserve previous internal granularity.
- npl_portfolio/property_collateral.py:
  Added `type_of_property_legacy_subtype` field.
- npl_portfolio/non_property_collateral_choices.py:
  Realigned COLLATERAL_TYPE_CHOICES (4.02) to the 14 EBA options and
  added TYPE_OF_APPRAISAL_AMOUNT_CHOICES.
- npl_portfolio/non_property_collateral.py:
  Added missing mandatory/recommended fields (4.21, 4.23-4.26) for
  external valuation and internal appraisal amounts.
- npl_portfolio/enforcement.py:
  Changed `number_of_failed_auctions` (4.42) from FloatField to IntegerField.
- npl_portfolio/admin.py:
  Updated PropertyCollateralAdmin and NonPropertyCollateralAdmin fieldsets
  to expose all newly added fields and the legacy subtype.

+++ npl_portfolio/enforcement.py
@@ -100,8 +100,8 @@ class Enforcement(models.Model):
     court_auction_reserve_price_for_last_auction = models.FloatField(blank=True, null=True,
                                                                      help_text=recommended_help('4.41', "Court set reserve price for last auction — minimum price required by the court."))
 
-    number_of_failed_auctions = models.FloatField(blank=True, null=True,
-                                                  help_text=recommended_help('4.42', "Number of failed previous auctions for the collateral. Applicable when 'yes' is selected in 'Enforcement status'."))
+    number_of_failed_auctions = models.IntegerField(blank=True, null=True,
+                                                    help_text=recommended_help('4.42', "Number of failed previous auctions for the collateral. Applicable when 'yes' is selected in 'Enforcement status'."))
 
     #
     # LEGACY DATA PROPERTIES (pre-2023 EBA draft — not in EU 2023/2083)
